Refactor license rendering: Remove CustomRenderer and introduce KafkaConnectorLicenseRenderer with separate TxtRenderer and CsvRenderer classes for improved structure and maintainability.
